requirements.txt
matplotlib==2.2.3
pandas==0.23.4
psycopg2-binary==2.7.5
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
